refactor(mvt): name Zod schema entry point
