+++ modules/mvt/test/index.ts
@@ -14,7 +14,7 @@ import './lib/vector-tiler/simplify-path.spec';
 import './lib/utils/geometry-utils.spec';
 
 import './map-style-loader.spec';
-import './tilejson-schema.spec';
+import './tilejson-zod-schema.spec';
 import './tilejson-loader.spec';
 
 import './mvt-loader.spec';
